+++ ArcaneWizardLibrary.lua
@@ -1,6 +1,11 @@
 local addonName, LIB = ...
 
+LIB.Internal = LIB.Internal or {}
+LIB.Internal.AddonChat = {}
+LIB.Internal.AddonLauncher = {}
+
 ArcaneWizardLibrary = ArcaneWizardLibrary or {}
+
 ArcaneWizardLibrary.Utils = {}
 ArcaneWizardLibrary.Dialogs = {}
 ArcaneWizardLibrary.Settings = {}